+++ R/mp_tmb_model_spec.R
@@ -7,8 +7,12 @@ TMBModelSpec = function(
     , must_save = character()
     , must_not_save = character()
     , sim_exprs = character()
+    , state_update = c("euler", "rk4", "euler_multinomial")
   ) {
   self = Base()
+  self$change_model = get_change_model(before, during, after)
+  self$state_update = get_state_update_type(match.arg(state_update), self$change_model)
+  self$update_method = get_state_update_method(self$state_update, self$change_model)
   self$before = before
   self$during = during
   self$after = after
@@ -18,7 +22,20 @@ TMBModelSpec = function(
   self$must_not_save = must_not_save
   self$sim_exprs = sim_exprs
   
-  self$expr_list = function() ExprList(self$before, self$during, self$after)
+  self$expr_list = function() {
+    ExprList(
+        self$update_method$before()
+      , self$update_method$during()
+      , self$update_method$after()
+    )
+  }
+  self$unrendered_expr_list = function() {
+    ExprList(
+        self$before
+      , self$during
+      , self$after
+    )
+  }
   
   self$all_derived_vars = function() {
     self$expr_list()$all_derived_vars()
@@ -71,6 +88,28 @@ TMBModelSpec = function(
         self$before, self$during, self$after
       , self$default, self$integers
       , self$must_save, self$must_not_save, self$sim_exprs
+      , self$state_update
+    )
+  }
+  self$change_update_method = function(state_update = c("euler", "rk4", "euler_multinomial")) {
+    if (self$state_update == "no") {
+      warning("This model has not formalized the notion of a state variable, and so changing how the state variables are updated has no effect.")
+    }
+    mp_tmb_model_spec(
+        self$before, self$during, self$after
+      , self$default, self$integers
+      , self$must_save, self$must_not_save, self$sim_exprs
+      , state_update
+    )
+  }
+  self$expand = function() {
+    mp_tmb_model_spec(
+        self$update_method$before()
+      , self$update_method$during()
+      , self$update_method$after()
+      , self$default, self$integers
+      , self$must_save, self$must_not_save, self$sim_exprs
+      , self$state_update
     )
   }
   self$tmb_model = function(
@@ -122,12 +161,13 @@ TMBModelSpec = function(
           )
         )
       )
-      , expr_list = ExprList(
-          before = self$before
-        , during = self$during 
-        , after = self$after
-        , .simulate_exprs = self$sim_exprs
-      )
+      , expr_list = self$expr_list()
+      #   ExprList(
+      #     before = self$before
+      #   , during = self$during 
+      #   , after = self$after
+      #   , .simulate_exprs = self$sim_exprs
+      # )
       , engine_methods = EngineMethods(
         int_vecs = do.call(IntVecs, self$all_integers())
       )
@@ -187,6 +227,9 @@ TMBModelSpec = function(
 #' being evaluated. For example, expressions that generate stochasticity should
 #' be listed in \code{sim_exprs} because TMB objective functions must be
 #' continuous.
+#' @param state_update (experimental) Optional character vector for how to update the
+#' state variables when it is relevant. Options include `"euler"`, `"rk4"`, 
+#' and `"euler_multinomial"`.
 #' @export
 mp_tmb_model_spec = TMBModelSpec
 
@@ -196,7 +239,8 @@ print.TMBModelSpec = function(x, ...) {
 }
 
 spec_printer = function(x, include_defaults) {
-  e = ExprList(x$before, x$during, x$after)
+  #e = ExprList(x$before, x$during, x$after)
+  #e = x$expr_list()
   if (include_defaults) {
     cat("---------------------\n")
     msg("Default values:\n") |> cat()
@@ -204,5 +248,7 @@ spec_printer = function(x, include_defaults) {
     print(melt_default_matrix_list(x$default), row.names = FALSE)
     cat("\n")
   }
-  print(e)
+  exprs = c(x$before, x$during, x$after)
+  schedule = c(length(x$before), length(x$during), length(x$after))
+  model_steps_printer(exprs, schedule)
 }