+++ R/tmb_model.R
@@ -151,7 +151,13 @@ TMBModel = function(
       params = self$params$vector(),
       random = self$random$vector()
     )
-    if (length(p$params) == 0L) p$params = 0
+    
+    ## FIXME: need a dummy parameter if the model has not
+    ## yet been parameterized. is there a more TMB-ish
+    ## way to do this?
+    if (length(p$params) == 0L) {
+      p$params = 0
+    } 
     p
   }
   self$random_arg = function() {
@@ -170,13 +176,12 @@ TMBModel = function(
       , random = self$random_arg()
       , DLL = tmb_cpp
       , silent = !verbose
-      #, type = getOption("macpan2_tmb_type")
-      #, checkParameterOrder = isTRUE(getOption("macpan2_tmb_check"))
     )
   }
   self$ad_fun = function(
         tmb_cpp = getOption("macpan2_dll")
       , verbose = getOption("macpan2_verbose")
+      , derivs = getOption("macpan2_tmb_derivs")
     ) {
     do.call(TMB::MakeADFun, self$make_ad_fun_arg(tmb_cpp))
   }
@@ -361,14 +366,17 @@ mp_trajectory_ensemble.TMBCalibrator = function(model, n, probs = c(0.025, 0.975
 ## not ready to export yet because we are not sure how to construct a single
 ## ad_fun that works both with process error simulation and deterministic 
 ## trajectory matching
-## 
+
+
 ##' Random Trajectory Simulations
 ##' 
-##' @param model Model object that can generate simulations with random numbers.
 ##' @param n Number of random trajectories to simulate.
 ##' @param probs Numeric vector of probabilities corresponding to quantiles for 
 ##' summarizing the results over the random realizations.
 ##' 
+##' @describeIn mp_trajectory Generate quantiles over `n` realizations of 
+##' the trajectory. Instead of a `value` column in the output data frame, there
+##' is one column for each of the quantiles defined in `probs`.
 ##' @export
 mp_trajectory_sim = function(model, n, probs = c(0.025, 0.25, 0.5, 0.75, 0.975)) {
   UseMethod("mp_trajectory_sim")
@@ -387,6 +395,7 @@ mp_trajectory_sim.TMBSimulator = function(model, n, probs = c(0.025, 0.25, 0.5, 
   cbind(r, rr)
 }
 
+##' @export
 mp_trajectory_sim.TMBCalibrator = function(model, n, probs = c(0.025, 0.25, 0.5, 0.75, 0.975)) {
   stop("Under construction")
 }
@@ -481,6 +490,8 @@ TMBSimulationUtils = function() {
   self$.runner = function(...
       , .phases = "during"
       , .method = c("report", "simulate", "sdreport")
+      , .sort = TRUE
+      , .values_only = FALSE
   ) {
     .method = match.arg(.method)
     compute_sd = FALSE
@@ -503,8 +514,11 @@ TMBSimulationUtils = function() {
       )
     }
     if (compute_sd) r$values = cbind(r$values, self$sdreport()$sd)
+    if (.values_only) return(r$values)
     s = self$.simulation_formatter(r, .phases)
-    s = s[order(s$time), , drop = FALSE] ## TODO: move sorting by time to the c++ side
+    if (.sort) {
+      s = s[order(s$time), , drop = FALSE] ## TODO: move sorting by time to the c++ side
+    }
     reset_rownames(s)
   }
   return_object(self, "TMBSimulationFormatter")